+++ openbis/source/java/ch/ethz/sis/openbis/generic/server/xls/export/helper/XLSDataSetTypeExportHelper.java
@@ -1,5 +1,6 @@
 package ch.ethz.sis.openbis.generic.server.xls.export.helper;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
@@ -16,35 +17,50 @@ import ch.ethz.sis.openbis.generic.asapi.v3.dto.entitytype.id.EntityTypePermId;
 import ch.ethz.sis.openbis.generic.asapi.v3.dto.entitytype.id.IEntityTypeId;
 import ch.ethz.sis.openbis.generic.asapi.v3.dto.plugin.Plugin;
 import ch.ethz.sis.openbis.generic.asapi.v3.dto.property.fetchoptions.PropertyAssignmentFetchOptions;
+import ch.ethz.sis.openbis.generic.server.xls.export.ExportableKind;
 import ch.ethz.sis.openbis.generic.server.xls.export.XLSExport;
 
 public class XLSDataSetTypeExportHelper extends AbstractXLSExportHelper
 {
 
+    public XLSDataSetTypeExportHelper(final Workbook wb)
+    {
+        super(wb);
+    }
+
     @Override
-    public int add(final IApplicationServerApi api, final String sessionToken, final Workbook wb,
-            final Collection<String> permIds, int rowNumber, final Map<String, Collection<String>> entityTypeExportPropertiesMap, final XLSExport.TextFormatting textFormatting)
+    public AdditionResult add(final IApplicationServerApi api, final String sessionToken, final Workbook wb,
+            final Collection<String> permIds, int rowNumber,
+            final Map<String, Collection<String>> entityTypeExportPropertiesMap,
+            final XLSExport.TextFormatting textFormatting)
     {
         assert permIds.size() == 1;
         final DataSetType dataSetType = getDataSetType(api, sessionToken, permIds.iterator().next());
+        final Collection<String> warnings = new ArrayList<>();
 
         if (dataSetType != null)
         {
-
-            addRow(wb, rowNumber++, true, "DATA_SET_TYPE");
-            addRow(wb, rowNumber++, true, "Version", "Code", "Validation script");
+            final String permId = dataSetType.getPermId().getPermId();
+            warnings.addAll(addRow(rowNumber++, true, ExportableKind.DATASET_TYPE, permId, "DATA_SET_TYPE"));
+            warnings.addAll(addRow(rowNumber++, true, ExportableKind.DATASET_TYPE, permId, "Version", "Code",
+                    "Validation script"));
 
             final Plugin validationPlugin = dataSetType.getValidationPlugin();
             final String script = validationPlugin != null
                     ? (validationPlugin.getName() != null ? validationPlugin.getName() + ".py" : "") : "";
 
-            addRow(wb, rowNumber++, false, "1", dataSetType.getCode(), script);
+            warnings.addAll(addRow(rowNumber++, false, ExportableKind.DATASET_TYPE, permId, "1",
+                    dataSetType.getCode(), script));
 
-            rowNumber = addEntityTypePropertyAssignments(wb, rowNumber, dataSetType.getPropertyAssignments());
-            return rowNumber + 1;
+            final AdditionResult additionResult = addEntityTypePropertyAssignments(rowNumber,
+                    dataSetType.getPropertyAssignments(), ExportableKind.DATASET_TYPE, permId);
+            warnings.addAll(additionResult.getWarnings());
+
+            rowNumber = additionResult.getRowNumber();
+            return new AdditionResult(rowNumber + 1, warnings);
         } else
         {
-            return rowNumber;
+            return new AdditionResult(rowNumber, warnings);
         }
     }
 
@@ -54,6 +70,8 @@ public class XLSDataSetTypeExportHelper extends AbstractXLSExportHelper
         fetchOptions.withValidationPlugin().withScript();
         final PropertyAssignmentFetchOptions propertyAssignmentFetchOptions = fetchOptions.withPropertyAssignments();
         propertyAssignmentFetchOptions.withPropertyType().withVocabulary();
+        propertyAssignmentFetchOptions.withPropertyType().withSampleType();
+        propertyAssignmentFetchOptions.withPropertyType().withMaterialType();
         propertyAssignmentFetchOptions.withPlugin().withScript();
         final Map<IEntityTypeId, DataSetType> dataSetTypes = api.getDataSetTypes(sessionToken,
                 Collections.singletonList(new EntityTypePermId(permId, EntityKind.DATA_SET)), fetchOptions);