+++ openbis/source/java/ch/ethz/sis/openbis/generic/server/xls/importer/XLSImport.java
@@ -62,7 +62,7 @@ public class XLSImport
 
     private final ExperimentTypeImportHelper experimentTypeHelper;
 
-    private final DatasetTypeImportHelper datasetHelper;
+    private final DatasetTypeImportHelper datasetTypeHelper;
 
     private final SpaceImportHelper spaceHelper;
 
@@ -97,7 +97,7 @@ public class XLSImport
         this.vocabularyTermHelper = new VocabularyTermImportHelper(this.delayedExecutor, mode, options, versions);
         this.sampleTypeHelper = new SampleTypeImportHelper(this.delayedExecutor, mode, options, versions);
         this.experimentTypeHelper = new ExperimentTypeImportHelper(this.delayedExecutor, mode, options, versions);
-        this.datasetHelper = new DatasetTypeImportHelper(this.delayedExecutor, mode, options, versions);
+        this.datasetTypeHelper = new DatasetTypeImportHelper(this.delayedExecutor, mode, options, versions);
         this.spaceHelper = new SpaceImportHelper(this.delayedExecutor, mode, options);
         this.projectHelper = new ProjectImportHelper(this.delayedExecutor, mode, options);
         this.experimentHelper = new ExperimentImportHelper(this.delayedExecutor, mode, options);
@@ -142,6 +142,51 @@ public class XLSImport
     }
 
     public List<IObjectId> importXLS(byte[] xls)
+    public ImportZipResult importZip(byte[] zip) throws IOException
+    {
+        final Map<String, String> scripts = new HashMap<>();
+        List<IObjectId> ids = null;
+        try
+        (
+                final ByteArrayInputStream bais = new ByteArrayInputStream(zip);
+                final ZipInputStream zis = new ZipInputStream(bais);
+                final BufferedInputStream bis = new BufferedInputStream(zis);
+        )
+        {
+            while (zis.available() > 0)
+            {
+                final ZipEntry zipEntry = zis.getNextEntry();
+                if (zipEntry != null)
+                {
+                    final String entryName = zipEntry.getName();
+                    if (entryName.startsWith("scripts/"))
+                    {
+                        final String scriptFileName = entryName.substring(8);
+                        final String scriptContent = new String(bis.readAllBytes());
+                        scripts.put(scriptFileName, scriptContent);
+                    } else if (entryName.endsWith(".xlsx"))
+                    {
+                        if (ids == null)
+                        {
+                            ids = importXLS(bis.readAllBytes());
+                        } else
+                        {
+                            throw new IllegalArgumentException("More than one XLSX file found in the archive.");
+                        }
+                    }
+                    zis.closeEntry();
+                }
+            }
+        }
+
+        if (ids == null)
+        {
+            throw new IllegalArgumentException("No XLSX file found in the archive.");
+        }
+        return new ImportZipResult(ids, scripts);
+    }
+
+    public List<IObjectId> importXLS(byte[] xls)
     {
         this.dbChecker.checkVersionsOnDataBase();
 
@@ -181,10 +226,11 @@ public class XLSImport
                         // parse and create scripts
                         scriptHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2, ScriptTypes.VALIDATION_SCRIPT);
                         // parse and create sample type
+                        boolean sTisNewVersion = sampleTypeHelper.isNewVersion(page, pageNumber, lineNumber, lineNumber + 2);
                         sampleTypeHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2);
                         semanticAnnotationImportHelper.importBlockForEntityType(page, pageNumber, lineNumber, lineNumber + 2, ImportTypes.SAMPLE_TYPE);
                         // parse and assignment properties
-                        if (lineNumber + 2 != blockEnd)
+                        if (sTisNewVersion && lineNumber + 2 != blockEnd)
                         {
                             scriptHelper.importBlock(page, pageNumber, lineNumber + 2, blockEnd, ScriptTypes.DYNAMIC_SCRIPT);
                             propertyHelper.importBlock(page, pageNumber, lineNumber + 2, blockEnd);
@@ -196,10 +242,11 @@ public class XLSImport
                         // parse and create scripts
                         scriptHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2, ScriptTypes.VALIDATION_SCRIPT);
                         // parse and create experiment type
+                        boolean eTisNewVersion = experimentTypeHelper.isNewVersion(page, pageNumber, lineNumber, lineNumber + 2);
                         experimentTypeHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2);
                         semanticAnnotationImportHelper.importBlockForEntityType(page, pageNumber, lineNumber, lineNumber + 2, ImportTypes.EXPERIMENT_TYPE);
                         // parse and assignment properties
-                        if (lineNumber + 2 != blockEnd)
+                        if (eTisNewVersion && lineNumber + 2 != blockEnd)
                         {
                             scriptHelper.importBlock(page, pageNumber, lineNumber + 2, blockEnd, ScriptTypes.DYNAMIC_SCRIPT);
                             propertyHelper.importBlock(page, pageNumber, lineNumber + 2, blockEnd);
@@ -211,10 +258,11 @@ public class XLSImport
                         // parse and create scripts
                         scriptHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2, ScriptTypes.VALIDATION_SCRIPT);
                         // parse and create dataset type
-                        datasetHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2);
+                        boolean dTisNewVersion = datasetTypeHelper.isNewVersion(page, pageNumber, lineNumber, lineNumber + 2);
+                        datasetTypeHelper.importBlock(page, pageNumber, lineNumber, lineNumber + 2);
                         semanticAnnotationImportHelper.importBlockForEntityType(page, pageNumber, lineNumber, lineNumber + 2, ImportTypes.DATASET_TYPE);
                         // parse and assignment properties
-                        if (lineNumber + 2 != blockEnd)
+                        if (dTisNewVersion && lineNumber + 2 != blockEnd)
                         {
                             scriptHelper.importBlock(page, pageNumber, lineNumber + 2, blockEnd, ScriptTypes.DYNAMIC_SCRIPT);
                             propertyHelper.importBlock(page, pageNumber, lineNumber + 2, blockEnd);