+++ pybis/src/python/pybis/__init__.py
@@ -1,7 +1,7 @@
 name = "pybis"
 __author__ = "Swen Vermeul • ID SIS • ETH Zürich"
 __email__ = "swen@ethz.ch"
-__version__ = "1.34.0"
+__version__ = "1.34.5"
 
 from . import pybis
 from .pybis import DataSet